+++ .sequelizerc
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve('config', 'config.js'),
-  'models-path': path.resolve('models'),
-  'migrations-path': path.resolve('migrations')
+  'config': path.resolve('backend', 'database', 'config', 'config.js'),
+  'models-path': path.resolve('backend', 'database', 'models'),
+  'migrations-path': path.resolve('backend', 'database', 'migrations')
 }
